refactor: reorganize tests to mirror src layout
- One test file per module: test_basis.py, test_design.py, test_experiment.py
- tests/regression/ mirrors src/pyneuroglm/regression/
- Shared MATLAB fixtures as pytest fixtures in tests/regression/conftest.py
- Parametrize empirical Bayes tests (6 -> 2 with @pytest.mark.parametrize)
- Merge cache hit/miss tests; remove namedtuple stdlib test
- Remove unused imports and verbose print statements
- Delete 5 old test files, add 9 new files
- 26 passed, 11 skipped (same coverage)
